Resolved unexpected behavior in .go export command. Now it exports a selected go to a .sql file in your server directory. Moved GetSpellIDFromLink to Chat.cpp.

+++ src/world/Objects/GameObject.cpp
@@ -247,7 +247,7 @@ void GameObject::SaveToFile(std::stringstream & name)
         << GetParentRotation(1) << ","
         << GetParentRotation(2) << ","
         << GetParentRotation(3) << ","
-        << GetState() << ","
+        << uint32(GetState()) << ","
         << GetFlags() << ","
         << GetFaction() << ","
         << GetScale() << ","
